rename test to match class, added unit test step for load calibration command
