+++ prototype2/gdgem/GdGemBaseTest.cpp
@@ -8,7 +8,7 @@
 #include <common/DataSave.h>
 #include <gdgem/GdGemBase.h>
 #include <../prototype2/adc_readout/test/TestUDPServer.h>
-#include <gdgem/GdGemReadoutTestData.h>
+#include <gdgem/GdGemBaseTestData.h>
 #include <test/TestBase.h>
 
 class GdGemBaseStandIn : public GdGemBase {
@@ -37,6 +37,25 @@ TEST_F(GdGemBaseTest, Constructor) {
   GdGemBaseStandIn Readout(Settings, LocalSettings);
   EXPECT_EQ(Readout.mystats.rx_packets, 0);
 }
+
+TEST_F(GdGemBaseTest, GetCalibrationCmd) {
+  GdGemBaseStandIn Readout(Settings, LocalSettings);
+  const int OutputSize = 1000;
+  unsigned int OutputBytes = 0;
+  char Output[OutputSize];
+  std::vector<std::string> badcmdlen = {"NMX_GET_CALIB"};
+  int res = Readout.getCalibration(badcmdlen, Output, &OutputBytes);
+  ASSERT_TRUE(res < 0);
+
+  std::vector<std::string> badcmdarg = {"NMX_GET_CALIB", "700", "0", "0"};
+  res = Readout.getCalibration(badcmdarg, Output, &OutputBytes);
+  ASSERT_TRUE(res < 0);
+
+  std::vector<std::string> goodcmd = {"NMX_GET_CALIB", "0", "0", "0"};
+  res = Readout.getCalibration(goodcmd, Output, &OutputBytes);
+  ASSERT_EQ(res, 0);
+}
+
 
 TEST_F(GdGemBaseTest, DataReceive) {
   GdGemBaseStandIn Readout(Settings, LocalSettings);
