+++ prototype2/gdgem/vmm3/CalibrationFile.cpp
@@ -101,11 +101,12 @@ bool CalibrationFile::addCalibration(unsigned int fecId, unsigned int vmmId,
 CalibrationFile::Calibration &
 CalibrationFile::getCalibration(unsigned int fecId, unsigned int vmmId,
                                 unsigned int chNo) {
+
   if ((fecId >= MAX_FEC) or (vmmId >= MAX_VMM) or (chNo >= MAX_CH)) {
     XTRACE(INIT, DEB, "invalid offsets: fec: %d, vmm: %d, ch:%d\n", fecId,
            vmmId, chNo);
     return ErrCorr;
   }
-
+  printf("XXXXXXXXXX\n");
   return Calibrations[fecId][vmmId][chNo];
 }
